P2-7: ProgressCursor.lastDocProcessed int→long, start at 0
Change ProgressCursor.lastDocProcessed from int to long to support
shards with >2B documents. Propagate the type change through:
- WorkItemCursor.progressCheckpointNum (int→long)
- WorkItem.startingDocId (Integer→Long)
- MigrationPipeline.migrateShard offset tracking (int→long)
- DocumentSource/LuceneSnapshotSource/SyntheticDocumentSource (int→long)
- PipelineDocumentsRunner startingOffset (int→long)

Change initial work item startingDocId from Integer.MIN_VALUE to 0,
removing the negative sentinel value.

+++ RFS/src/main/java/org/opensearch/migrations/bulkload/pipeline/PipelineDocumentsRunner.java
@@ -81,7 +81,7 @@ public class PipelineDocumentsRunner {
                     log.info("Pipeline acquired work item: {}", wi);
 
                     var shardId = new ShardId(snapshotName, wi.getIndexName(), wi.getShardNumber());
-                    int startingOffset = wi.getStartingDocId() != null && wi.getStartingDocId() >= 0
+                    long startingOffset = wi.getStartingDocId() != null && wi.getStartingDocId() >= 0
                         ? wi.getStartingDocId() : 0;
 
                     var pipeline = new MigrationPipeline(source, sink, maxDocsPerBatch, maxBytesPerBatch);
